+++ gala/coordinates/reflex.py
@@ -1,5 +1,4 @@
 import astropy.coordinates as coord
-from .galactocentric import get_galactocentric2019
 
 __all__ = ['reflex_correct']
 
@@ -33,7 +32,7 @@ def reflex_correct(coords, galactocentric_frame=None):
 
     # If not specified, use the Astropy default Galactocentric frame
     if galactocentric_frame is None:
-        galactocentric_frame = get_galactocentric2019()
+        galactocentric_frame = coord.Galactocentric()
 
     v_sun = galactocentric_frame.galcen_v_sun