+++ gala/coordinates/velocity_frame_transforms.py
@@ -2,8 +2,6 @@
 
 import astropy.coordinates as coord
 
-from .galactocentric import get_galactocentric2019
-
 __all__ = ["vgsr_to_vhel", "vhel_to_vgsr"]
 
 
@@ -39,7 +37,7 @@ def vgsr_to_vhel(coordinate, vgsr, vsun=None):
     """
 
     if vsun is None:
-        galcen = get_galactocentric2019()
+        galcen = coord.Galactocentric()
         vsun = galcen.galcen_v_sun.to_cartesian().xyz
 
     return vgsr - _get_vproj(coordinate, vsun)
@@ -70,7 +68,7 @@ def vhel_to_vgsr(coordinate, vhel, vsun):
     """
 
     if vsun is None:
-        galcen = get_galactocentric2019()
+        galcen = coord.Galactocentric()
         vsun = galcen.galcen_v_sun.to_cartesian().xyz
 
     return vhel + _get_vproj(coordinate, vsun)